refactor to prevent dupl
