+++ datafusion/physical-plan/src/joins/nested_loop_join.rs
@@ -778,45 +778,11 @@ impl ExecutionPlan for NestedLoopJoinExec {
         let left = ctx.encode_child(self.left())?;
         let right = ctx.encode_child(self.right())?;
 
-        let join_type = match self.join_type() {
-            JoinType::Inner => protobuf::JoinType::Inner,
-            JoinType::Left => protobuf::JoinType::Left,
-            JoinType::Right => protobuf::JoinType::Right,
-            JoinType::Full => protobuf::JoinType::Full,
-            JoinType::LeftSemi => protobuf::JoinType::Leftsemi,
-            JoinType::RightSemi => protobuf::JoinType::Rightsemi,
-            JoinType::LeftAnti => protobuf::JoinType::Leftanti,
-            JoinType::RightAnti => protobuf::JoinType::Rightanti,
-            JoinType::LeftMark => protobuf::JoinType::Leftmark,
-            JoinType::RightMark => protobuf::JoinType::Rightmark,
-        };
+        let join_type = crate::joins::proto::join_type_to_proto(*self.join_type());
 
         let filter = self
             .filter()
-            .map(|f| -> Result<protobuf::JoinFilter> {
-                let expression = ctx.encode_expr(f.expression())?;
-                let column_indices = f
-                    .column_indices()
-                    .iter()
-                    .map(|i| {
-                        let side = match i.side {
-                            JoinSide::Left => protobuf::JoinSide::LeftSide,
-                            JoinSide::Right => protobuf::JoinSide::RightSide,
-                            JoinSide::None => protobuf::JoinSide::None,
-                        };
-                        protobuf::ColumnIndex {
-                            index: i.index as u32,
-                            side: side.into(),
-                        }
-                    })
-                    .collect();
-                let schema = f.schema().as_ref().try_into()?;
-                Ok(protobuf::JoinFilter {
-                    expression: Some(expression),
-                    column_indices,
-                    schema: Some(schema),
-                })
-            })
+            .map(|f| crate::joins::proto::join_filter_to_proto(f, ctx))
             .transpose()?;
 
         Ok(Some(protobuf::PhysicalPlanNode {
@@ -864,70 +830,16 @@ impl NestedLoopJoinExec {
             "right",
         )?;
 
-        let join_type =
-            match protobuf::JoinType::try_from(join.join_type).map_err(|_| {
-                internal_datafusion_err!(
-                    "NestedLoopJoinExec: unknown JoinType {}",
-                    join.join_type
-                )
-            })? {
-                protobuf::JoinType::Inner => JoinType::Inner,
-                protobuf::JoinType::Left => JoinType::Left,
-                protobuf::JoinType::Right => JoinType::Right,
-                protobuf::JoinType::Full => JoinType::Full,
-                protobuf::JoinType::Leftsemi => JoinType::LeftSemi,
-                protobuf::JoinType::Rightsemi => JoinType::RightSemi,
-                protobuf::JoinType::Leftanti => JoinType::LeftAnti,
-                protobuf::JoinType::Rightanti => JoinType::RightAnti,
-                protobuf::JoinType::Leftmark => JoinType::LeftMark,
-                protobuf::JoinType::Rightmark => JoinType::RightMark,
-            };
+        let join_type = crate::joins::proto::join_type_from_proto(
+            join.join_type,
+            "NestedLoopJoinExec",
+        )?;
 
         let filter = join
             .filter
             .as_ref()
-            .map(|f| -> Result<JoinFilter> {
-                let schema: Schema = f
-                    .schema
-                    .as_ref()
-                    .ok_or_else(|| {
-                        internal_datafusion_err!(
-                            "NestedLoopJoinExec: JoinFilter missing schema"
-                        )
-                    })?
-                    .try_into()?;
-                let expression = ctx.decode_required_expr(
-                    f.expression.as_ref(),
-                    &schema,
-                    "NestedLoopJoinExec",
-                    "filter.expression",
-                )?;
-                let column_indices = f
-                    .column_indices
-                    .iter()
-                    .map(|i| {
-                        let side =
-                            match protobuf::JoinSide::try_from(i.side).map_err(|_| {
-                                internal_datafusion_err!(
-                                    "NestedLoopJoinExec: unknown JoinSide {}",
-                                    i.side
-                                )
-                            })? {
-                                protobuf::JoinSide::LeftSide => JoinSide::Left,
-                                protobuf::JoinSide::RightSide => JoinSide::Right,
-                                protobuf::JoinSide::None => JoinSide::None,
-                            };
-                        Ok(ColumnIndex {
-                            index: i.index as usize,
-                            side,
-                        })
-                    })
-                    .collect::<Result<Vec<_>>>()?;
-                Ok(JoinFilter::new(
-                    expression,
-                    column_indices,
-                    Arc::new(schema),
-                ))
+            .map(|f| {
+                crate::joins::proto::join_filter_from_proto(f, ctx, "NestedLoopJoinExec")
             })
             .transpose()?;
